requirements.txt
streamlit
python-dotenv
pinecone-client==5.*
sentence-transformers
pypdf
python-pptx
ollama
google-generativeai>=0.3.0
